pyproject.toml
[project]
name = "mcdonalds-etl"
version = "0.1.0"
description = "ETL Process for Mcdonalds"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "adlfs",
    "dagster==1.10.*",
    "dagster-webserver",
    "dagster-cloud",
    "dagster-duckdb",
    "pandas",
    "polars-lts-cpu",
    "pyarrow",
    "pyodbc",
    "python-dotenv",
    "SQLAlchemy"
]

[tool.uv]
dev-dependencies = [
    "ruff",
    "pytest"
]

[tool.dagster]
module_name = "mcdonalds_etl.definitions"
code_location_name = "mcdonalds_etl"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mcdonalds_etl"]